+++ tests/sentry/tasks/test_process_buffer.py
@@ -8,7 +8,6 @@ from sentry.tasks.process_buffer import (
     process_incr,
     process_pending,
     process_pending_batch,
-    schedule_delayed_workflows,
 )
 from sentry.testutils.cases import TestCase
 
@@ -51,71 +50,3 @@ class ProcessPendingBatchTest(TestCase):
         with self.assertNoLogs("sentry.tasks.process_buffer", level="WARNING"):
             process_pending_batch()
             assert len(mock_process_buffer.mock_calls) == 1
-
-
-class ScheduleDelayedWorkflowsTest(TestCase):
-    @mock.patch("sentry.options.get")
-    @mock.patch("sentry.rules.processing.buffer_processing.process_buffer_for_type")
-    @mock.patch("sentry.workflow_engine.tasks.delayed_workflows.DelayedWorkflow")
-    def test_schedule_delayed_workflows_locked_out(
-        self,
-        mock_delayed_workflow: mock.MagicMock,
-        mock_process_buffer_for_type: mock.MagicMock,
-        mock_options_get: mock.MagicMock,
-    ) -> None:
-        # Mock the config option to return True (using new scheduling task)
-        mock_options_get.return_value = True
-
-        with self.assertLogs("sentry.tasks.process_buffer", level="WARNING") as logger:
-            lock = get_process_lock("schedule_delayed_workflows")
-            with lock.acquire():
-                schedule_delayed_workflows()
-                self.assertEqual(len(logger.output), 1)
-                assert len(mock_process_buffer_for_type.mock_calls) == 0
-
-        with self.assertNoLogs("sentry.tasks.process_buffer", level="WARNING"):
-            schedule_delayed_workflows()
-            assert len(mock_process_buffer_for_type.mock_calls) == 1
-            mock_process_buffer_for_type.assert_called_with(
-                "delayed_workflow", mock_delayed_workflow
-            )
-
-    @mock.patch("sentry.options.get")
-    @mock.patch("sentry.rules.processing.buffer_processing.process_buffer_for_type")
-    @mock.patch("sentry.workflow_engine.tasks.delayed_workflows.DelayedWorkflow")
-    def test_schedule_delayed_workflows_config_option_true(
-        self,
-        mock_delayed_workflow: mock.MagicMock,
-        mock_process_buffer_for_type: mock.MagicMock,
-        mock_options_get: mock.MagicMock,
-    ) -> None:
-        # Mock the config option to return False (not using new scheduling task)
-        mock_options_get.return_value = False
-
-        with self.assertLogs("sentry.tasks.process_buffer", level="INFO") as logger:
-            schedule_delayed_workflows()
-            # Should log and exit without calling process_buffer_for_type
-            assert len(mock_process_buffer_for_type.mock_calls) == 0
-            assert any(
-                "Configured to use process_pending_batch" in output for output in logger.output
-            )
-
-    @mock.patch("sentry.options.get")
-    @mock.patch("sentry.rules.processing.buffer_processing.process_buffer_for_type")
-    @mock.patch("sentry.workflow_engine.tasks.delayed_workflows.DelayedWorkflow")
-    def test_schedule_delayed_workflows_normal_operation(
-        self,
-        mock_delayed_workflow: mock.MagicMock,
-        mock_process_buffer_for_type: mock.MagicMock,
-        mock_options_get: mock.MagicMock,
-    ) -> None:
-        # Mock the config option to return True (using new scheduling task)
-        mock_options_get.return_value = True
-
-        schedule_delayed_workflows()
-
-        # Should only process delayed_workflow
-        assert len(mock_process_buffer_for_type.mock_calls) == 1
-        mock_process_buffer_for_type.assert_called_once_with(
-            "delayed_workflow", mock_delayed_workflow
-        )
